Simulation::StartSimulation complete
StartSimulation now complete and returns either a time taken to complete simulation as an integer value or the value of 0 if the simulation fails.

Time is a spurious value which is actually the total number of itterations through the drive loop.

Also created Simulation::DriveSector and completed this to simulate the driving of the vehicle. Includes std::sleep_for to slow down processing for some form of simulated realism. Suggest comment out std::sleep_for if number crunching.

Huge changes to calibration.cpp too to add extra methods, private variables etc that were required for the simulation to work. My favourite being the function Calibration::GetTorque as it was great fun coding an interpolation algorithm.